resin-proxy-config: Add recipe redsocks configuration
We add a resin-proxy-config service that starts a script. This script looks for
a system-proxy folder in the boot partition that should have a [redsocks.conf
file](https://github.com/darkk/redsocks/blob/master/redsocks.conf.example).

If this file is present, it creates iptables rules to redirect all tcp traffic
to port 12345 except local and docker connections.

This means that the redsocks.conf has to configure redsocks to listen on port
12345 (with the local_port option) and set `daemon = off` (because of systemd
service configuration) to work properly.

The recipe also allows using a no_proxy file inside /mnt/boot/system-proxy to
define IPs/subnets that will not be routed through the proxy.

Signed-off-by: Pablo Carranza Velez <pablo@resin.io>
Signed-off-by: Andrei Gherzan <andrei@resin.io>