prepare-openvpn: Switch to watching for config.json changes in order to update the VPN config
The watching works by having the config-json.path trigger the config-json.service every time the config.json file changes, and as the prepare-openvpn.service has a `WantedBy=config-json.service` it will therefore also be triggered by it.  As config-json.service also has a `WantedBy=mnt-boot.mount` it means it will also trigger once on boot, and hence trigger the prepare-openvpn.service once on boot as well, allowing for the case of subsequent boots.